Extract createSocket and make configurable
